backend/requirements.txt
fastapi
uvicorn[standard]
motor
PyPDF2
passlib
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
pydantic-settings
python-dotenv
pydantic
email-validator
